requirements.txt
Selenium==4.17.2
python-decouple==3.8
requests==2.31.0
selenium-requests==2.0.3